requirements.txt
Flask~=3.1.0
Flask-SQLAlchemy~=3.1.1
SQLAlchemy~=2.0.36
pandas~=2.2.3
waitress~=3.0.2
onnxruntime~=1.20.1
numpy<2.0
# numpy~=2.2.1
pillow~=11.0.0
PyYAML~=6.0.2
pytest~=8.3.4
scikit-learn~=1.6.0  # for test data
torchvision~=0.17.2  # for test data
requests-mock~=1.12.1  # for testing
pytest-mock~=3.14.0  # for testing
requests~=2.32.3
psycopg2-binary~=2.9.10
